refactor(tools): extract provider schema formatters into tools/schema.rs
Move the provider-specific tool-schema formatters out of the 3,034-line
tools/mod.rs into a focused submodule:

- tools/schema.rs (212) — params_to_json_schema, resolve_params, and the
                          public tools_openai / tools_anthropic / tools_google

mod.rs drops to 2,830 lines. The three formatters are re-exported so the
`tools::tools_*` call sites in gateway/providers/* are unchanged. tools/mod.rs
remains large because it is predominantly a flat ToolDef registry (all_tools()
plus per-category definition blocks), which would need a registry redesign to
split further — out of scope here.

cargo clippy --workspace --all-targets -- -D warnings: clean
cargo fmt --all -- --check: clean
